requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
fastapi==0.109.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
joblib==1.3.2
numpy==1.26.3
openai==1.11.0
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.12.0
sniffio==1.3.0
starlette==0.35.1
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
